BinaryParsingTools

Documentation for BinaryParsingTools.

BinaryParsingTools.Tag

BinaryParsingTools.cache_stream_in_ram

BinaryParsingTools.read_at

BinaryParsingTools.swap_endianness

BinaryParsingTools.@serializable

BinaryParsingTools.Tag — Type

N-string tag.

BinaryParsingTools.cache_stream_in_ram — Function

cache_stream_in_ram(io::IO, ::Type{T})

Determine whether to cache the full IOStream in RAM or not to increase performance, using an IOBuffer.

Defaults to true. This is advised for small-ish files, because access will be much easier if only a single IO access is required (one for reading the whole binary blob). On the other hand, that will increase RAM consumption, and dramatically so for large files.

Additional parameters may be specified by returning a value of type Tuple{Bool, NamedTuple}, where the meaning of the Bool is unchanged and NamedTuple describes parameters for IOBuffer, such as maxsize.

BinaryParsingTools.read_at — Method

Read a value of type T located at an offset from a given start (defaulting to the current position), without modifying the stream position.

BinaryParsingTools.swap_endianness — Function

swap_endianness(io::IO, ::Type{T})

Decide whether to swap the endianness on any data read from the provided IO (defaults to false if not extended).

If the binary format being used chooses a specific endianness, you may extend this function for the related type you are trying to parse and decide depending on host endianness.

For example, assuming the data is in little-endian (1) or big-endian (2) format:

# The host is in little-endian format if `ENDIAN_BOM == 0x04030201` (see the documentation for `Base.ENDIAN_BOM`).
# If that is the case then assuming binary data is expected in little-endian format there is no need to swap the endianness.
swap_endianness(io::IO, ::Type{BinaryData}) = ENDIAN_BOM == 0x01020304 # data comes in as little-endian, host is big-endian

# That will be the opposite if the binary data format is expected in big-endian; little-endian machines
# will need to swap the endianness.
swap_endianness(io::IO, ::Type{BinaryData}) = ENDIAN_BOM == 0x04030201 # data comes in as big-endian, host is little-endian

If you don't know in which endianness the data comes, you would typically parse a magic number to determine whether endianness differs from the host:

swap_endianness(io::IO, ::Type{BinaryData}) = read(io, UInt32) == 0x87654321 # expecting 0x12345678 as magic number

BinaryParsingTools.@serializable — Macro

Mark a given struct as serializable, automatically implementing Base.read.

If some of the structure members are vectors, their length must be specified using a syntax of the form params::Vector{UInt32} => param_count where param_count can be any expression, which may depend on other structure members.

Fields can be read in a custom manner by using a syntax of the form params::SomeField << ex where ex can be e.g. read(io, SomeField, other_field.length) where other_field can refer to any previous field in the struct. This expression may refer to a special variable __origin__, which is the position of the IO before parsing the struct.

Additional arguments required for Base.read can be specified with the syntax @arg name at the very start of the structure, before any actual fields. In this way, the definition for Base.read will include these extra arguments. Calling code will then have to provide these extra arguments.

If a struct field is reserved in a binary format, and may be safely ignored when reading from IO, use @reserved <nbytes>. In this way, the field will not be part of the struct, and instead <nbytes> bytes will be skipped at that point during parsing.

LineNumberNodes will be preserved and inserted wherever necessary to keep stack traces informative.

Examples

@serializable struct MarkArrayTable
  mark_count::UInt16
  mark_records::Vector{MarkRecord} => mark_count
end

@serializable struct LigatureAttachTable
  @arg mark_class_count # will need to be provided when `Base.read`ing this type.

  # Length of `component_records`.
  component_count::UInt16

  component_records::Vector{Vector{UInt16}} << [[read(io, UInt16) for _ in 1:mark_class_count] for _ in 1:component_count]
end

Here is an advanced example which makes use of all the features:

@serializable struct LigatureArrayTable
  @arg mark_class_count # will need to be provided when `Base.read`ing this type.

  # Length of `ligature_attach_offsets`.
  ligature_count::UInt16

  # Declare a reserved field, usually meant for padding.
  # No field will be added to the struct itself, but 6 bytes will
  # be skipped at that point before parsing the next field.
  @reserved 6

  # Offsets in bytes from the origin of the structure to data blocks formatted as `LigatureAttachTable`s.
  ligature_attach_offsets::Vector{UInt16} => ligature_count

  ligature_attach_tables::Vector{LigatureAttachTable} << [read_at(io, LigatureAttachTable, offset, mark_class_count; start = __origin__) for offset in ligature_attach_offsets]
end